# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses/sessions/{courseSession}/quiz-questions": {
            "get": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Get quiz questions for a specific course session",
                "description": "Retrieve all quiz questions associated with a specific course session, including media and answers.",
                "operationId": "getQuizQuestions",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "ID of the course session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuestionResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course session not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/quizzes": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get a list of quizzes",
                "description": "Display a listing of the resource.",
                "operationId": "eb9ba15b35ba53ab91ba8b1a4e215e29",
                "parameters": [
                    {
                        "name": "subjectId",
                        "in": "query",
                        "description": "The ID of the subject to filter quizzes by",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userSaved",
                        "in": "query",
                        "description": "Filter quizzes saved by the authenticated user",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of quizzes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuizResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/quizzes/subjects": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get a list of quiz subjects",
                "description": "Get quiz subjects.",
                "operationId": "4a2d43fe13afd36d916f7ccf4f92ed7e",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Filter subjects by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of quiz subjects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubjectResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/quizzes/{quiz}/questions": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get a list of questions for a quiz",
                "description": "Get quiz questions.",
                "operationId": "f1652f20b49e9b566c3745044fd820fb",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "The ID of the quiz",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of quiz questions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuestionResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Quiz not found"
                    }
                }
            }
        },
        "/api/quizzes/{quiz}/save": {
            "post": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Add a course to the user's saved list",
                "description": "Add a course to the authenticated user's saved list.",
                "operationId": "75bb549d2aca8950dee8b719d8921adb",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "The ID of the quiz to save",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Quiz successfully added to saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Quiz not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/quizzes/{quiz}/unsave": {
            "delete": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Remove a course from the user's saved list",
                "description": "Remove a course from the authenticated user's saved list.",
                "operationId": "72dfad6dcc323e043e4df3e88ff0f952",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "The ID of the quiz to unsave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Quiz successfully removed from saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Quiz not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/students/quizzes": {
            "get": {
                "tags": [
                    "Student Quizzes"
                ],
                "summary": "Get a list of student quizzes",
                "operationId": "7228b2729a0df30c7d0ce141e27ce91e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentQuizResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Student Quizzes"
                ],
                "summary": "Store a new student quiz",
                "operationId": "765f9e4a6874c3a7b45e3cfda18b9b4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreStudentQuizRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quiz stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentQuizResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreStudentQuizRequest": {
                "required": [
                    "quizId",
                    "quizType",
                    "solvedQuestions",
                    "timeTaked"
                ],
                "properties": {
                    "quizId": {
                        "description": "The ID of the quiz",
                        "type": "integer"
                    },
                    "quizType": {
                        "description": "The type of the quiz",
                        "type": "string",
                        "enum": [
                            "quiz",
                            "courseSession"
                        ]
                    },
                    "solvedQuestions": {
                        "description": "The number of questions solved",
                        "type": "integer"
                    },
                    "timeTaked": {
                        "description": "The time taken to complete the quiz",
                        "type": "string",
                        "format": "minute:seconds"
                    }
                },
                "type": "object"
            },
...
            "AnswerResource": {
                "title": "Answer Resource",
                "description": "Answer resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the answer",
                        "type": "integer"
                    },
                    "content": {
                        "description": "The content of the answer",
                        "type": "string"
                    },
                    "correct": {
                        "description": "Indicates if the answer is correct",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "CourseSessionResource": {
                "title": "Course Session Resource",
                "description": "Course Session resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the course session",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the course session",
                        "type": "string"
                    },
                    "note": {
                        "description": "Additional notes for the course session",
                        "type": "string"
                    },
                    "time": {
                        "description": "The time of the course session",
                        "type": "string"
                    },
                    "like": {
                        "description": "The number of likes for the course session",
                        "type": "integer"
                    },
                    "dislike": {
                        "description": "The number of dislikes for the course session",
                        "type": "integer"
                    },
                    "type": {
                        "description": "The type of the course session",
                        "type": "string"
                    },
                    "isFree": {
                        "description": "Indicates if the course session is free",
                        "type": "boolean"
                    },
                    "teacher": {
                        "$ref": "#/components/schemas/TeacherResource"
                    },
                    "course": {
                        "$ref": "#/components/schemas/CourseResource"
                    },
                    "video": {
                        "description": "The video media associated with the course session",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    },
                    "image": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "currentUserProgress": {
                        "$ref": "#/components/schemas/UserProgressResource"
                    }
                },
                "type": "object"
            },
            "MediaResource": {
                "title": "MediaResource",
                "description": "Media resource representation for files and images",
                "required": [
                    "id",
                    "name",
                    "fileName",
                    "collection",
                    "url",
                    "size",
                    "type",
                    "extension",
                    "caption"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the media",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the media file without extension",
                        "type": "string",
                        "example": "nature_photo"
                    },
                    "fileName": {
                        "description": "Complete file name with extension",
                        "type": "string",
                        "example": "nature_photo.jpg"
                    },
                    "collection": {
                        "description": "Collection name where the media is stored",
                        "type": "string",
                        "example": "photos"
                    },
                    "url": {
                        "description": "Full URL to access the media file",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/media/nature_photo.jpg"
                    },
                    "size": {
                        "description": "Human-readable file size",
                        "type": "string",
                        "example": "2.5 MB"
                    },
                    "type": {
                        "description": "Type of media",
                        "type": "string",
                        "enum": [
                            "image",
                            "video",
                            "document",
                            "audio",
                            "other"
                        ],
                        "example": "image"
                    },
                    "extension": {
                        "description": "File extension",
                        "type": "string",
                        "example": "jpg"
                    },
                    "caption": {
                        "description": "Caption or description of the media. Falls back to name if not set",
                        "type": "string",
                        "example": "Beautiful nature photograph"
                    },
                    "width": {
                        "description": "Width of the image in pixels (only for images)",
                        "type": "integer",
                        "format": "int32",
                        "example": 1920,
                        "nullable": true
                    },
                    "height": {
                        "description": "Height of the image in pixels (only for images)",
                        "type": "integer",
                        "format": "int32",
                        "example": 1080,
                        "nullable": true
                    },
                    "thumbnailUrl": {
                        "description": "URL of the thumbnail version. Only present if 'thumb' conversion exists",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/media/conversions/nature_photo-thumb.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "QuestionResource": {
                "title": "Question Resource",
                "description": "Question resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the question",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the question",
                        "type": "string"
                    },
                    "type": {
                        "description": "The type of the question",
                        "type": "string"
                    },
                    "media": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "answers": {
                        "description": "The answers associated with the question",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnswerResource"
                        }
                    },
                    "note": {
                        "description": "Additional notes or comments for the question",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "QuizResource": {
                "title": "Quiz Resource",
                "description": "Quiz resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the quiz",
                        "type": "integer"
                    },
                    "subject": {
                        "$ref": "#/components/schemas/SubjectResource"
                    },
                    "title": {
                        "description": "The title of the quiz",
                        "type": "string"
                    },
                    "time": {
                        "description": "The time allocated for the quiz in minutes",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "StudentQuizResource": {
                "properties": {
                    "id": {
                        "description": "The unique identifier of the student quiz",
                        "type": "integer"
                    },
                    "quiz": {
                        "$ref": "#/components/schemas/QuizResource"
                    },
                    "courseQuiz": {
                        "$ref": "#/components/schemas/CourseSessionResource"
                    },
                    "solvedQuestions": {
                        "description": "The number of questions solved by the student",
                        "type": "integer"
                    },
                    "totalQuestions": {
                        "description": "The total number of questions in the quiz",
                        "type": "integer"
                    },
                    "timeTaked": {
                        "description": "The time taken by the student to complete the quiz",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "SubjectResource": {
                "title": "Subject Resource",
                "description": "Subject resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the subject",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the subject",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the subject",
                        "type": "string"
                    },
                    "year": {
                        "description": "The academic year of the subject",
                        "type": "integer"
                    },
                    "semester": {
                        "description": "The semester in which the subject is taught",
                        "type": "integer"
                    },
                    "university": {
                        "description": "The university associated with the subject",
                        "type": "object",
                        "nullable": true
                    },
                    "quizezCounts": {
                        "description": "The count of quizzes associated with the subject",
                        "type": "integer",
                        "nullable": true
                    },
                    "questionsCount": {
                        "description": "The count of questions associated with the subject",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}